Address review feedback from jortel.
- Remove `ok` return from EndSessionURL; caller checks logoutURL != ""
- Remove defensive nil check on rpClient (RpClient returns valid or error)
- Use direct err assignment instead of shadow variables (pErr, sErr, eErr)
- Distinguish NotFound from operational errors in findSubject
- Remove nil check on idpHandler (Storage always constructed with one)
- Move tests from idp_test.go to pkg_test.go per hub convention
- Build end_session URL via url.Parse (preserves existing query params)
